Task 1/requirements.txt
pinecone-client
langchain
langchain-google-genai
langchain-pinecone
pypdf
python-dotenv